1. Print out the test name for each test 2. Add a new contructor (colNames, cols) for BasicaTable. Fix the bug of tableName when writeing table to db. 3. Add new test for uploading table and left join 2. 4. Fix the bug of isNull in BasicShort

+++ src/com/xxdb/data/BasicShort.java
@@ -23,7 +23,7 @@ public class BasicShort extends AbstractScalar{
 	
 	@Override
 	public boolean isNull() {
-		return  value == Integer.MIN_VALUE;
+		return  value == Short.MIN_VALUE;
 	}
 	
 	@Override
